Add Resource::hash()
