+++ node-graph/libraries/resources/Cargo.toml
@@ -16,7 +16,6 @@ dyn-any = { workspace = true }
 core-types = { workspace = true }
 
 # Workspace dependencies
-blake3 = { workspace = true }
 url = { workspace = true }
 
 # Optional workspace dependencies
